move Q specific functions to Q package
